requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0,<2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0,<4.0
transformers>=4.36.0,<5.0
torch
ollama>=0.1.0
openai>=1.10.0
scikit-learn>=1.3.0
torchvision>=0.15.0